Move get_machine_name to a more public place
As this function is used outside of ert, it should not be hidden below
ensemble_evalutor.config.

port_handler.py serves as a placeholder for the code, but the function
but exposed and should be used as ert.shared.get_machine.name()
