+++ src/ert/shared/__init__.py
@@ -7,5 +7,6 @@ except ImportError:
 from ert.shared.hook_implementations.jobs import (
     _resolve_ert_share_path as ert_share_path,
 )
+from ert.shared.port_handler import get_machine_name
 
-__all__ = ["ert_share_path", "__version__"]
+__all__ = ["ert_share_path", "get_machine_name", "__version__"]
